Added stopwatch icon and visualisations
Frame adding works like that - Press enter then fill the prompt. After that you have to click on the object and it will generate keyframe then press num 1. Move it and press 0. After this point you have to click with the mouse on the object to generate keyframe and then press num 1 to collect it. After witch 0 will write it in the console.